api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
minio==7.2.12
pdfplumber==0.11.4
python-docx==1.1.2
httpx==0.28.1
qdrant-client==1.13.3
fpdf2==2.8.2
